+++ strawberry/types/arguments.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import inspect
 import warnings
-from collections.abc import Iterable, Mapping
 from typing import (
     TYPE_CHECKING,
     Annotated,
@@ -18,18 +17,21 @@ from strawberry.exceptions import MultipleStrawberryArgumentsError, UnsupportedT
 from strawberry.scalars import is_scalar
 from strawberry.types.base import (
     StrawberryList,
+    StrawberryMaybe,
     StrawberryOptional,
     has_object_definition,
 )
 from strawberry.types.enum import EnumDefinition
 from strawberry.types.lazy_type import LazyType, StrawberryLazyReference
+from strawberry.types.maybe import Some
 from strawberry.types.unset import UNSET as _deprecated_UNSET  # noqa: N811
 from strawberry.types.unset import (
-    Some,
     _deprecated_is_unset,  # noqa: F401
 )
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable, Mapping
+
     from strawberry.schema.config import StrawberryConfig
     from strawberry.types.base import StrawberryType
     from strawberry.types.scalar import ScalarDefinition, ScalarWrapper
@@ -137,8 +139,9 @@ class StrawberryArgument:
 
         return is_graphql_generic(self.type)
 
+    @property
     def is_maybe(self) -> bool:
-        return isinstance(self.type, StrawberryOptional) and self.type.is_maybe
+        return isinstance(self.type, StrawberryMaybe)
 
 
 def convert_argument(
@@ -150,7 +153,8 @@ def convert_argument(
     # TODO: move this somewhere else and make it first class
     if isinstance(type_, StrawberryOptional):
         res = convert_argument(value, type_.of_type, scalar_registry, config)
-        return Some(res) if type_.is_maybe else res
+
+        return Some(res) if isinstance(type_, StrawberryMaybe) else res
 
     if value is None:
         return None
@@ -159,7 +163,7 @@ def convert_argument(
         return _deprecated_UNSET
 
     if isinstance(type_, StrawberryList):
-        value_list = cast(Iterable, value)
+        value_list = cast("Iterable", value)
         return [
             convert_argument(x, type_.of_type, scalar_registry, config)
             for x in value_list
@@ -183,7 +187,7 @@ def convert_argument(
 
         type_definition = type_.__strawberry_definition__
         for field in type_definition.fields:
-            value = cast(Mapping, value)
+            value = cast("Mapping", value)
             graphql_name = config.name_converter.from_field(field)
 
             if graphql_name in value:
@@ -194,7 +198,7 @@ def convert_argument(
                     config,
                 )
 
-        type_ = cast(type, type_)
+        type_ = cast("type", type_)
         return type_(**kwargs)
 
     raise UnsupportedTypeError(type_)
