+++ strawberry/types/object_type.py
@@ -20,7 +20,7 @@ from strawberry.exceptions import (
     ObjectIsNotClassError,
 )
 from strawberry.types.base import get_object_definition
-from strawberry.types.unset import _annotation_is_maybe
+from strawberry.types.maybe import _annotation_is_maybe
 from strawberry.utils.deprecations import DEPRECATION_MESSAGES, DeprecatedDescriptor
 from strawberry.utils.str_converters import to_camel_case
 
@@ -127,8 +127,8 @@ def _inject_default_for_maybe_annotations(
     cls: builtins.type[T], annotations: dict[str, Any]
 ) -> None:
     """Inject `= None` for fields with `Maybe` annotations and no default value."""
-    for name, annot in annotations.copy().items():
-        if _annotation_is_maybe(annot) and not hasattr(cls, name):
+    for name, annotation in annotations.copy().items():
+        if _annotation_is_maybe(annotation) and not hasattr(cls, name):
             setattr(cls, name, None)
